datasets/e_stat/pyproject.toml
[project]
name = "queria-estat"
version = "0.1.0"
description = "e-Stat dataset for queria"
requires-python = ">=3.13"
dependencies = [
    "queria",
    "dlt[ducklake]>=1.17.0",
    "estat-api-dlt-helper>=0.3.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ingestion"]

[dependency-groups]
dev = ["pytest>=9.0.2"]

[tool.uv.sources]
queria = { path = "../../", editable = true }
